live input works

+++ synth/synthMessages.ts
@@ -54,7 +54,7 @@ export interface SendSharedArrayBuffers extends Message {
     flag: MessageFlag.sharedArrayBuffers,
     livePitches: Int8Array,
     bassLivePitches: Int8Array,
-    liveInputValues: Int8Array,
+    liveInputValues: Uint32Array,
     livePitchInstruments: Int8Array,
     liveBassPitchInstruments: Int8Array
 }
